+++ pinot-common/src/main/java/org/apache/pinot/common/function/scalar/ArithmeticFunctions.java
@@ -28,8 +28,6 @@ import org.apache.pinot.spi.annotations.ScalarFunction;
  */
 public class ArithmeticFunctions {
 
-  private static final double DOUBLE_COMPARISON_TOLERANCE = 1e-7d;
-
   private ArithmeticFunctions() {
   }
 
@@ -155,21 +153,5 @@ public class ArithmeticFunctions {
   @ScalarFunction
   public static double truncate(double a) {
     return Math.signum(a) * Math.floor(Math.abs(a));
-  }
-
-  @ScalarFunction(names = {"gte"})
-  public static boolean greaterThanOrEquals(double a, double b) {
-    return a >= b;
-  }
-
-  @ScalarFunction(names = {"lte"})
-  public static boolean lessThanOrEquals(double a, double b) {
-    return a <= b;
-  }
-
-  @ScalarFunction
-  public static boolean equals(double a, double b) {
-    // To avoid approximation errors
-    return (Math.abs(a - b) < DOUBLE_COMPARISON_TOLERANCE);
   }
 }
